Extract sign-up handlers to separate module
