+++ src/handlers/index.js
@@ -3,4 +3,5 @@ module.exports = {
     ...require('./handlers'),
     ...require('./tournaments.handlers'),
     ...require('./users.handlers'),
+    ...require('./signUp.handlers'),
 };
